move method to its own file
