+++ packages/core/lib/commands/migrate/index.js
@@ -1,5 +1,6 @@
 module.exports = {
   run: require("./run"),
   meta: require("./meta"),
-  determineDryRunSettings: require("./determineDryRunSettings")
+  determineDryRunSettings: require("./determineDryRunSettings"),
+  prepareConfigForRealMigrations: require("./prepareConfigForRealMigrations")
 };
